Add files via upload
This C++ application prompts the user to input an integer. The program then converts the integer into a string representation, reverses this string, and checks if the reversed version matches the original string. If the two strings match, the inputted number is a palindrome. The result, whether the number is a palindrome or not, is then displayed to the user.